# speed up cast of Row types