Summary:
The CastExpr when applied to Row Types does a lot of copies of strings and shared_ptrs, as well as unnecessary construction of vectors that slow it down.  This is especially apparent when casting a Row with a lot of children and the cast only changes the field names.

This diff improves the speed of casting rows by over 5x in some cases by using references instead of copying, and reusing the toType instead of constructing a new type on every invocation.

Added a benchmark for casting

Before
```
============================================================================
[...]pression/benchmarks/CastBenchmark.cpp     relative  time/iter   iters/s
============================================================================
scalar                                                     10.11ns    98.96M
renameSmallStruct                                           6.72ns   148.89M
renameLargeStruct                                         758.64ns     1.32M
smallStructNested                                          26.29ns    38.04M
largeStructNested                                           9.24us   108.17K
```

After:
```
============================================================================
[...]pression/benchmarks/CastBenchmark.cpp     relative  time/iter   iters/s
============================================================================
scalar                                                      9.10ns   109.92M
renameSmallStruct                                           4.57ns   219.01M
renameLargeStruct                                         143.23ns     6.98M
smallStructNested                                          24.84ns    40.25M
largeStructNested                                           8.37us   119.41K
```

Differential Revision: D41741945